refactor: Remove duplicate filter token component (#6013)
