+++ packages/hooks/src/index.ts
@@ -9,4 +9,5 @@ export * from './useIsMounted'
 export { usePreloadImages } from './usePreloadImage'
 export { default as useIntersectionObserver } from './useIntersectionObserver'
 export { default as useDelayedUnmount } from './useDelayedUnmount'
+export { default as useSortedTokensByQuery } from './useSortedTokensByQuery'
 export { default as useTheme, COOKIE_THEME_KEY, THEME_DOMAIN } from './useTheme'
