+++ packages/utils/package.json
@@ -7,6 +7,7 @@
   "private": true,
   "dependencies": {
     "@pancakeswap/tokens": "*",
-    "@pancakeswap/sdk": "*"
+    "@pancakeswap/swap-sdk-core": "*",
+    "@pancakeswap/token-lists": "*"
   }
 }
